requirements.txt
numpy
scipy
matplotlib
seaborn
tqdm
python-doctr[torch]
torch
fastapi[standard]
pillow
uvicorn
python-multipart